pyproject.toml
[tool.poetry]
name = "fastapi-auth-user"
version = "0.1.0"
description = "auth user"
authors = ["Vittalius <Vittalius@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "fastapi_auth_user"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.95.0"
uvicorn = "^0.21.1"
sqlalchemy = "^2.0.8"
python-dotenv = "^1.0.0"
psycopg2 = "^2.9.7"
pyjwt = "^2.6.0"
python-jose = "^3.3.0"
passlib = "^1.7.4"
bcrypt = "^4.0.1"
python-decouple = "^3.8"
alembic = "^1.12.0"
psycopg2-binary = "^2.9.7"
python-multipart = "^0.0.6"
jinja2 = "^3.1.2"

[tool.poetry.scripts]
start = "fastapi-auth-usert.src.__main__:app"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
